requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
tiktoken==0.7.0
openai==1.43.0
tenacity==8.5.0
numpy==1.26.4
pandas==2.2.2
# Removed heavy native deps for broader compatibility
uvicorn==0.30.5
fastapi==0.112.2
jinja2==3.1.4
pyyaml==6.0.2
requests==2.32.3
tree_sitter==0.21.3
javalang==0.13.0
beautifulsoup4==4.12.3
pytest==8.3.2
pytest-cov==5.0.0
rich==13.7.1
click==8.1.7
aiofiles==24.1.0
python-slugify==8.0.4
packaging==24.1
networkx==3.3

